requirements.txt
fastapi[standard]
google-generativeai
dotenv